chore: split PlaylistsPage gallery out of the CRUD page
Move PlaylistGallery and Virtuoso/dnd wrappers next to PlaylistCard so the page stays list/CRUD only, without changing queryKey or AI filter behavior.
