+++ docs/roadmap.md
@@ -172,6 +172,7 @@ Both P0 rows (#1–#2) are closed — the full v17 audit pack landed (after one 
 
 | Branch | Status | Notes |
 |--------|--------|-------|
+| `chore/split-playlists-page` | 🔄 in progress | Mechanical split: `PlaylistsPage` list/CRUD; `PlaylistGallery` + virtuoso wrappers under `components/playlists/` |
 | `fix/artist-autocomplete-throttle-cancellation` | 🔄 started | Add Artist second-pass: `user` throttle + Main abort of superseded waves; intervals unchanged |
 | `feat/add-artist-autocomplete-artist-only` | ✅ merged | [#150](https://github.com/KazeKaze93/RuleDesk/pull/150) — Add Artist autocomplete: Gelbooru `category===artist`; Rule34 top-5 DAPI second-pass; Browse search unfiltered |
 | `feat/sync-status-live-ui` | ✅ merged | [#149](https://github.com/KazeKaze93/RuleDesk/pull/149) — invalidate `["artists"]` on `sync:artist` + repair start/end; preload wires `REPAIR_*`; `sync:progress` payload unchanged |
